Move to variant comparator
