+++ src/function/scalar/create_sort_key.cpp
@@ -9,21 +9,12 @@
 
 #include "duckdb/common/enums/order_type.hpp"
 #include "duckdb/common/radix.hpp"
-#include "duckdb/common/helper.hpp"
-#include "duckdb/common/types/datetime.hpp"
 #include "duckdb/common/types/interval.hpp"
-#include "duckdb/common/types/uhugeint.hpp"
-#include "duckdb/common/types/bignum.hpp"
-#include "duckdb/common/types/variant.hpp"
-#include "duckdb/common/enum_util.hpp"
-#include "duckdb/function/scalar/variant_utils.hpp"
 #include "duckdb/function/scalar/generic_functions.hpp"
 #include "duckdb/execution/expression_executor.hpp"
 #include "duckdb/planner/expression/bound_function_expression.hpp"
 #include "duckdb/planner/expression_binder.hpp"
 #include "duckdb/parser/parser.hpp"
-
-#include <algorithm>
 
 namespace duckdb {
 
@@ -419,442 +410,6 @@ struct SortKeyBlobOperator {
 		return pos + 1;
 	}
 };
-
-//===--------------------------------------------------------------------===//
-// Variant Encoding
-//===--------------------------------------------------------------------===//
-// VARIANT is physically a STRUCT (keys/children/values/data), but ordering it by that physical
-// representation does not match the logical ordering of the values it stores. Instead we traverse
-// the variant and emit a sort key per value as [type-rank][value]. Sorting is performed by type
-// first and value second. Object keys are processed in string-sorted order so that objects that
-// only differ in key order (e.g. {'b': 10, 'a': 1} and {'a': 1, 'b': 10}) compare equal.
-//
-// Types are grouped into "ranks" the way a semi-structured system (e.g. Snowflake) groups them:
-// all integers, decimals and bignums fold into a single NUMBER rank (compared by numeric value) and
-// FLOAT/DOUBLE fold into a single REAL rank. Everything else keeps its own rank.
-
-//! The sort-key rank of a variant value - determines the cross-type ordering (type-first)
-enum class VariantSortRank : data_t {
-	NULL_VALUE = 0,
-	BOOLEAN,
-	NUMBER,
-	REAL,
-	VARCHAR,
-	BLOB,
-	UUID,
-	DATE,
-	TIME_MICROS,
-	TIME_NANOS,
-	TIME_MICROS_TZ,
-	TIMESTAMP_SEC,
-	TIMESTAMP_MILIS,
-	TIMESTAMP_MICROS,
-	TIMESTAMP_NANOS,
-	TIMESTAMP_MICROS_TZ,
-	TIMESTAMP_NANOS_TZ,
-	INTERVAL,
-	GEOMETRY,
-	BITSTRING,
-	ARRAY,
-	OBJECT
-};
-
-static data_t GetVariantTypeRank(VariantLogicalType type_id) {
-	switch (type_id) {
-	case VariantLogicalType::VARIANT_NULL:
-		return static_cast<data_t>(VariantSortRank::NULL_VALUE);
-	case VariantLogicalType::BOOL_TRUE:
-	case VariantLogicalType::BOOL_FALSE:
-		return static_cast<data_t>(VariantSortRank::BOOLEAN);
-	// all integers, decimals and bignums fold into a single NUMBER rank (compared by numeric value)
-	case VariantLogicalType::INT8:
-	case VariantLogicalType::INT16:
-	case VariantLogicalType::INT32:
-	case VariantLogicalType::INT64:
-	case VariantLogicalType::INT128:
-	case VariantLogicalType::UINT8:
-	case VariantLogicalType::UINT16:
-	case VariantLogicalType::UINT32:
-	case VariantLogicalType::UINT64:
-	case VariantLogicalType::UINT128:
-	case VariantLogicalType::DECIMAL:
-	case VariantLogicalType::BIGNUM:
-		return static_cast<data_t>(VariantSortRank::NUMBER);
-	// FLOAT and DOUBLE fold into a single REAL rank
-	case VariantLogicalType::FLOAT:
-	case VariantLogicalType::DOUBLE:
-		return static_cast<data_t>(VariantSortRank::REAL);
-	case VariantLogicalType::VARCHAR:
-		return static_cast<data_t>(VariantSortRank::VARCHAR);
-	case VariantLogicalType::BLOB:
-		return static_cast<data_t>(VariantSortRank::BLOB);
-	case VariantLogicalType::UUID:
-		return static_cast<data_t>(VariantSortRank::UUID);
-	case VariantLogicalType::DATE:
-		return static_cast<data_t>(VariantSortRank::DATE);
-	case VariantLogicalType::TIME_MICROS:
-		return static_cast<data_t>(VariantSortRank::TIME_MICROS);
-	case VariantLogicalType::TIME_NANOS:
-		return static_cast<data_t>(VariantSortRank::TIME_NANOS);
-	case VariantLogicalType::TIME_MICROS_TZ:
-		return static_cast<data_t>(VariantSortRank::TIME_MICROS_TZ);
-	case VariantLogicalType::TIMESTAMP_SEC:
-		return static_cast<data_t>(VariantSortRank::TIMESTAMP_SEC);
-	case VariantLogicalType::TIMESTAMP_MILIS:
-		return static_cast<data_t>(VariantSortRank::TIMESTAMP_MILIS);
-	case VariantLogicalType::TIMESTAMP_MICROS:
-		return static_cast<data_t>(VariantSortRank::TIMESTAMP_MICROS);
-	case VariantLogicalType::TIMESTAMP_NANOS:
-		return static_cast<data_t>(VariantSortRank::TIMESTAMP_NANOS);
-	case VariantLogicalType::TIMESTAMP_MICROS_TZ:
-		return static_cast<data_t>(VariantSortRank::TIMESTAMP_MICROS_TZ);
-	case VariantLogicalType::TIMESTAMP_NANOS_TZ:
-		return static_cast<data_t>(VariantSortRank::TIMESTAMP_NANOS_TZ);
-	case VariantLogicalType::INTERVAL:
-		return static_cast<data_t>(VariantSortRank::INTERVAL);
-	case VariantLogicalType::GEOMETRY:
-		return static_cast<data_t>(VariantSortRank::GEOMETRY);
-	case VariantLogicalType::BITSTRING:
-		return static_cast<data_t>(VariantSortRank::BITSTRING);
-	case VariantLogicalType::ARRAY:
-		return static_cast<data_t>(VariantSortRank::ARRAY);
-	case VariantLogicalType::OBJECT:
-		return static_cast<data_t>(VariantSortRank::OBJECT);
-	default:
-		throw NotImplementedException("Variant type %s is not supported in create_sort_key",
-		                              EnumUtil::ToString(type_id));
-	}
-}
-
-//===--------------------------------------------------------------------===//
-// NUMBER rank encoding
-//===--------------------------------------------------------------------===//
-// All integers, decimals and bignums fold into a single NUMBER rank and are compared by numeric
-// value. To make values of different scales/widths comparable (and equal when numerically equal),
-// each value is reduced to an order-preserving "decimal-scientific" form:
-//
-//     value = sign * 0.d1 d2 ... dk * 10^adjexp     (d1 != 0, trailing zeros stripped)
-//
-// and encoded as [class][adjexp][digits][terminator], where everything after the class byte is
-// complemented for negatives. This makes e.g. 1::TINYINT, 1::BIGINT, 1.00::DECIMAL and 1::BIGNUM all
-// encode identically, and orders them numerically (so -100.5 < 0 < 1.5 < 15 < ...).
-struct VariantNumberKey {
-	//! 0 = negative, 1 = zero, 2 = positive (so negatives sort before zero before positives)
-	data_t number_class;
-	int64_t adjusted_exponent;
-	//! significant decimal digits ('0'..'9'), leading digit non-zero, trailing zeros stripped
-	string digits;
-};
-
-//! Build a number key from a sign, the decimal digits of the magnitude (no leading zeros, as e.g.
-//! produced by Uhugeint::ToString) and a base-10 scale (value = magnitude / 10^scale)
-static VariantNumberKey BuildNumberKey(bool negative, const string &magnitude_digits, int64_t scale) {
-	VariantNumberKey key;
-	// find the first non-zero digit (magnitude strings have no leading zeros, but be defensive)
-	idx_t first = 0;
-	while (first < magnitude_digits.size() && magnitude_digits[first] == '0') {
-		first++;
-	}
-	if (first == magnitude_digits.size()) {
-		// the value is zero
-		key.number_class = 1;
-		key.adjusted_exponent = 0;
-		return key;
-	}
-	key.number_class = negative ? 0 : 2;
-	// position of the leading digit: (#digits after leading zeros - 1) - scale
-	key.adjusted_exponent = static_cast<int64_t>(magnitude_digits.size() - first) - 1 - scale;
-	// strip trailing zeros (keep at least the leading digit)
-	idx_t last = magnitude_digits.size();
-	while (last > first + 1 && magnitude_digits[last - 1] == '0') {
-		last--;
-	}
-	key.digits = magnitude_digits.substr(first, last - first);
-	return key;
-}
-
-template <class SINK>
-static void VariantEncodeNumber(SINK &sink, const VariantNumberKey &key) {
-	// the class byte is the primary discriminator and is never complemented
-	sink.Write(key.number_class);
-	if (key.number_class == 1) {
-		// zero - the class fully describes the value
-		return;
-	}
-	const bool negative = key.number_class == 0;
-	// for negatives we complement every payload byte so that larger magnitudes sort earlier
-	auto emit = [&](data_t b) {
-		sink.Write(negative ? static_cast<data_t>(~b) : b);
-	};
-
-	// adjusted exponent (order-preserving signed encoding) orders magnitude
-	data_t exp_buffer[sizeof(int64_t)];
-	Radix::EncodeData<int64_t>(exp_buffer, key.adjusted_exponent);
-	for (idx_t i = 0; i < sizeof(int64_t); i++) {
-		emit(exp_buffer[i]);
-	}
-	// significant digits (1..10 so the 0x00 terminator is smaller than any digit -> prefix sorts first)
-	for (auto c : key.digits) {
-		emit(static_cast<data_t>((c - '0') + 1));
-	}
-	emit(0); // terminator
-}
-
-//! Decompose a 128-bit signed value (hugeint) into (sign, magnitude)
-static uhugeint_t HugeintMagnitude(hugeint_t value, bool &negative) {
-	negative = value.upper < 0;
-	uint64_t hi = static_cast<uint64_t>(value.upper);
-	uint64_t lo = value.lower;
-	if (negative) {
-		// two's complement negate to get the magnitude
-		hi = ~hi;
-		lo = ~lo;
-		if (++lo == 0) {
-			++hi;
-		}
-	}
-	return uhugeint_t(hi, lo);
-}
-
-//! Decompose a variant integer value into (sign, 128-bit magnitude)
-static uhugeint_t VariantIntegralMagnitude(VariantLogicalType type_id, const_data_ptr_t ptr, bool &negative) {
-	negative = false;
-	switch (type_id) {
-	case VariantLogicalType::INT8:
-		return HugeintMagnitude(hugeint_t(Load<int8_t>(ptr)), negative);
-	case VariantLogicalType::INT16:
-		return HugeintMagnitude(hugeint_t(Load<int16_t>(ptr)), negative);
-	case VariantLogicalType::INT32:
-		return HugeintMagnitude(hugeint_t(Load<int32_t>(ptr)), negative);
-	case VariantLogicalType::INT64:
-		return HugeintMagnitude(hugeint_t(Load<int64_t>(ptr)), negative);
-	case VariantLogicalType::INT128:
-		return HugeintMagnitude(Load<hugeint_t>(ptr), negative);
-	case VariantLogicalType::UINT8:
-		return uhugeint_t(0, Load<uint8_t>(ptr));
-	case VariantLogicalType::UINT16:
-		return uhugeint_t(0, Load<uint16_t>(ptr));
-	case VariantLogicalType::UINT32:
-		return uhugeint_t(0, Load<uint32_t>(ptr));
-	case VariantLogicalType::UINT64:
-		return uhugeint_t(0, Load<uint64_t>(ptr));
-	case VariantLogicalType::UINT128:
-		return Load<uhugeint_t>(ptr);
-	default:
-		throw InternalException("VariantIntegralMagnitude called on non-integer type");
-	}
-}
-
-//! Compute the number key for any value in the NUMBER rank (integer, decimal or bignum)
-static VariantNumberKey VariantGetNumberKey(VariantLogicalType type_id, const UnifiedVariantVectorData &variant,
-                                            idx_t row, uint32_t values_idx, const_data_ptr_t value_ptr) {
-	switch (type_id) {
-	case VariantLogicalType::DECIMAL: {
-		auto decimal_data = VariantUtils::DecodeDecimalData(variant, row, values_idx);
-		hugeint_t unscaled;
-		switch (decimal_data.GetPhysicalType()) {
-		case PhysicalType::INT16:
-			unscaled = hugeint_t(Load<int16_t>(decimal_data.value_ptr));
-			break;
-		case PhysicalType::INT32:
-			unscaled = hugeint_t(Load<int32_t>(decimal_data.value_ptr));
-			break;
-		case PhysicalType::INT64:
-			unscaled = hugeint_t(Load<int64_t>(decimal_data.value_ptr));
-			break;
-		default:
-			unscaled = Load<hugeint_t>(decimal_data.value_ptr);
-			break;
-		}
-		bool negative;
-		auto magnitude = HugeintMagnitude(unscaled, negative);
-		return BuildNumberKey(negative, Uhugeint::ToString(magnitude), decimal_data.scale);
-	}
-	case VariantLogicalType::BIGNUM: {
-		auto bignum_blob = VariantUtils::DecodeStringData(variant, row, values_idx);
-		auto decimal_string = Bignum::BignumToVarchar(bignum_t(bignum_blob));
-		bool negative = !decimal_string.empty() && decimal_string[0] == '-';
-		idx_t offset = negative ? 1 : 0;
-		return BuildNumberKey(negative, decimal_string.substr(offset), 0);
-	}
-	default: {
-		// integer types
-		bool negative;
-		auto magnitude = VariantIntegralMagnitude(type_id, value_ptr, negative);
-		return BuildNumberKey(negative, Uhugeint::ToString(magnitude), 0);
-	}
-	}
-}
-
-//! Sink that only computes the encoded length of a variant value
-struct VariantSortKeyLengthSink {
-	idx_t length = 0;
-
-	inline void Write(data_t) {
-		length++;
-	}
-	inline void WriteBytes(const_data_ptr_t, idx_t count) {
-		length += count;
-	}
-};
-
-//! Sink that writes the encoded bytes of a variant value (flipping bytes for DESCENDING order)
-struct VariantSortKeyWriteSink {
-	VariantSortKeyWriteSink(data_ptr_t result_ptr, idx_t offset, bool flip_bytes)
-	    : result_ptr(result_ptr), offset(offset), flip_bytes(flip_bytes) {
-	}
-
-	data_ptr_t result_ptr;
-	idx_t offset;
-	bool flip_bytes;
-
-	inline void Write(data_t b) {
-		result_ptr[offset++] = flip_bytes ? static_cast<data_t>(~b) : b;
-	}
-	inline void WriteBytes(const_data_ptr_t src, idx_t count) {
-		for (idx_t i = 0; i < count; i++) {
-			Write(src[i]);
-		}
-	}
-};
-
-template <class T, class SINK>
-static void VariantEncodeFixed(SINK &sink, T value) {
-	data_t buffer[sizeof(T)];
-	Radix::EncodeData<T>(buffer, value);
-	sink.WriteBytes(buffer, sizeof(T));
-}
-
-template <class SINK>
-static void VariantEncodeString(SINK &sink, const string_t &str, bool is_varchar) {
-	auto input_data = const_data_ptr_cast(str.GetData());
-	auto input_size = str.GetSize();
-	if (is_varchar) {
-		// VARCHAR is valid UTF-8 (never contains 0xFF) so we use the +1 / null-delimiter scheme
-		for (idx_t i = 0; i < input_size; i++) {
-			sink.Write(static_cast<data_t>(input_data[i] + 1));
-		}
-	} else {
-		// arbitrary bytes - escape \x00 and \x01 so they cannot collide with the delimiter
-		for (idx_t i = 0; i < input_size; i++) {
-			if (input_data[i] <= 1) {
-				sink.Write(SortKeyVectorData::BLOB_ESCAPE_CHARACTER);
-			}
-			sink.Write(input_data[i]);
-		}
-	}
-	sink.Write(SortKeyVectorData::STRING_DELIMITER);
-}
-
-template <class SINK>
-static void EncodeVariantValue(const UnifiedVariantVectorData &variant, idx_t row, uint32_t values_idx, SINK &sink) {
-	auto type_id = variant.GetTypeId(row, values_idx);
-	// write the type rank - this guarantees values are ordered by type first
-	sink.Write(GetVariantTypeRank(type_id));
-
-	auto blob_ptr = const_data_ptr_cast(variant.GetData(row).GetData());
-	auto value_ptr = blob_ptr + variant.GetByteOffset(row, values_idx);
-	switch (type_id) {
-	case VariantLogicalType::VARIANT_NULL:
-		// no payload - the rank fully describes the value
-		break;
-	case VariantLogicalType::BOOL_TRUE:
-		sink.Write(1);
-		break;
-	case VariantLogicalType::BOOL_FALSE:
-		sink.Write(0);
-		break;
-	// all integers, decimals and bignums fold into the NUMBER rank and compare by numeric value
-	case VariantLogicalType::INT8:
-	case VariantLogicalType::INT16:
-	case VariantLogicalType::INT32:
-	case VariantLogicalType::INT64:
-	case VariantLogicalType::INT128:
-	case VariantLogicalType::UINT8:
-	case VariantLogicalType::UINT16:
-	case VariantLogicalType::UINT32:
-	case VariantLogicalType::UINT64:
-	case VariantLogicalType::UINT128:
-	case VariantLogicalType::DECIMAL:
-	case VariantLogicalType::BIGNUM:
-		VariantEncodeNumber(sink, VariantGetNumberKey(type_id, variant, row, values_idx, value_ptr));
-		break;
-	case VariantLogicalType::FLOAT:
-		// fold FLOAT into the REAL rank by widening to double (lossless)
-		VariantEncodeFixed<double>(sink, static_cast<double>(Load<float>(value_ptr)));
-		break;
-	case VariantLogicalType::DOUBLE:
-		VariantEncodeFixed<double>(sink, Load<double>(value_ptr));
-		break;
-	case VariantLogicalType::UUID:
-		VariantEncodeFixed<hugeint_t>(sink, Load<hugeint_t>(value_ptr));
-		break;
-	case VariantLogicalType::DATE:
-		VariantEncodeFixed<int32_t>(sink, Load<int32_t>(value_ptr));
-		break;
-	case VariantLogicalType::TIME_MICROS:
-	case VariantLogicalType::TIME_NANOS:
-	case VariantLogicalType::TIMESTAMP_SEC:
-	case VariantLogicalType::TIMESTAMP_MILIS:
-	case VariantLogicalType::TIMESTAMP_MICROS:
-	case VariantLogicalType::TIMESTAMP_NANOS:
-	case VariantLogicalType::TIMESTAMP_MICROS_TZ:
-	case VariantLogicalType::TIMESTAMP_NANOS_TZ:
-		// all of these are stored as a single (UTC-normalized) int64 that compares directly
-		VariantEncodeFixed<int64_t>(sink, Load<int64_t>(value_ptr));
-		break;
-	case VariantLogicalType::TIME_MICROS_TZ:
-		// TIME WITH TIME ZONE requires a dedicated byte-comparable transform
-		VariantEncodeFixed<uint64_t>(sink, Load<dtime_tz_t>(value_ptr).sort_key());
-		break;
-	case VariantLogicalType::INTERVAL:
-		// normalize the interval so that equal intervals encode identically
-		VariantEncodeFixed<interval_t>(sink, Load<interval_t>(value_ptr).Normalize());
-		break;
-	case VariantLogicalType::VARCHAR:
-		VariantEncodeString(sink, VariantUtils::DecodeStringData(variant, row, values_idx), true);
-		break;
-	case VariantLogicalType::BLOB:
-	case VariantLogicalType::GEOMETRY:
-	case VariantLogicalType::BITSTRING:
-		VariantEncodeString(sink, VariantUtils::DecodeStringData(variant, row, values_idx), false);
-		break;
-	case VariantLogicalType::ARRAY: {
-		auto nested_data = VariantUtils::DecodeNestedData(variant, row, values_idx);
-		for (idx_t i = 0; i < nested_data.child_count; i++) {
-			auto child_values_idx = variant.GetValuesIndex(row, nested_data.children_idx + i);
-			EncodeVariantValue(variant, row, child_values_idx, sink);
-		}
-		sink.Write(SortKeyVectorData::LIST_DELIMITER);
-		break;
-	}
-	case VariantLogicalType::OBJECT: {
-		auto nested_data = VariantUtils::DecodeNestedData(variant, row, values_idx);
-		// gather (key, value) pairs and process them in string-sorted key order so that objects
-		// that only differ in key order compare equal
-		vector<std::pair<string_t, uint32_t>> entries;
-		entries.reserve(nested_data.child_count);
-		for (idx_t i = 0; i < nested_data.child_count; i++) {
-			auto key_idx = variant.GetKeysIndex(row, nested_data.children_idx + i);
-			auto child_values_idx = variant.GetValuesIndex(row, nested_data.children_idx + i);
-			entries.emplace_back(variant.GetKey(row, key_idx), child_values_idx);
-		}
-		std::sort(entries.begin(), entries.end(),
-		          [](const std::pair<string_t, uint32_t> &a, const std::pair<string_t, uint32_t> &b) {
-			          return a.first < b.first;
-		          });
-		for (auto &entry : entries) {
-			VariantEncodeString(sink, entry.first, true);
-			EncodeVariantValue(variant, row, entry.second, sink);
-		}
-		sink.Write(SortKeyVectorData::LIST_DELIMITER);
-		break;
-	}
-	default:
-		throw NotImplementedException("Variant type %s is not supported in create_sort_key",
-		                              EnumUtil::ToString(type_id));
-	}
-}
 
 struct SortKeyListEntry {
 	static bool IsArray() {
@@ -1410,54 +965,6 @@ void CreateSortKeyHelpers::CreateSortKey(const Vector &input, OrderModifiers ord
 	sort_key_data.push_back(make_uniq<SortKeyVectorData>(input, input_count, order_modifier));
 
 	CreateSortKeyInternal(sort_key_data, modifiers, false, result, input_count);
-}
-
-void CreateSortKeyHelpers::CreateVariantSortKey(const Vector &input, idx_t count, OrderModifiers modifiers,
-                                                Vector &result) {
-	// VARIANT is physically a STRUCT, but the generic create_sort_key encodes that physical
-	// representation (which is reversible, so it can be used by aggregates to store/decode values).
-	// For comparison / ordering we instead encode the *logical* value of the variant - this encoding
-	// is intentionally not reversible (e.g. all integer widths fold together), so it lives here and is
-	// only used by the variant_comparator collation. NULLs are propagated into the result validity.
-	RecursiveUnifiedVectorFormat variant_format;
-	Vector::RecursiveToUnifiedFormat(input, variant_format);
-	UnifiedVariantVectorData variant(variant_format);
-
-	data_t null_byte = SortKeyVectorData::NULL_FIRST_BYTE;
-	data_t valid_byte = SortKeyVectorData::NULL_LAST_BYTE;
-	if (modifiers.null_type == OrderByNullType::NULLS_LAST) {
-		std::swap(null_byte, valid_byte);
-	}
-	const bool flip_bytes = modifiers.order_type == OrderType::DESCENDING;
-
-	result.SetVectorType(VectorType::FLAT_VECTOR);
-	auto result_data = FlatVector::GetDataMutable<string_t>(result);
-	auto &result_validity = FlatVector::ValidityMutable(result);
-
-	for (idx_t r = 0; r < count; r++) {
-		const bool valid = variant.RowIsValid(r);
-		// phase 1 - compute the encoded length
-		idx_t length = 1; // validity byte
-		if (valid) {
-			VariantSortKeyLengthSink length_sink;
-			EncodeVariantValue(variant, r, 0, length_sink);
-			length += length_sink.length;
-		}
-		// phase 2 - allocate and construct
-		result_data[r] = StringVector::EmptyString(result, length);
-		auto ptr = data_ptr_cast(result_data[r].GetDataWriteable());
-		if (!valid) {
-			ptr[0] = null_byte;
-			result_data[r].Finalize();
-			// propagate NULL so that NULL = NULL stays NULL and ORDER BY ... NULLS FIRST/LAST is honored
-			result_validity.SetInvalid(r);
-			continue;
-		}
-		ptr[0] = valid_byte;
-		VariantSortKeyWriteSink write_sink(ptr, 1, flip_bytes);
-		EncodeVariantValue(variant, r, 0, write_sink);
-		result_data[r].Finalize();
-	}
 }
 
 void CreateSortKeyHelpers::CreateSortKey(DataChunk &input, const vector<OrderModifiers> &modifiers, Vector &result) {
