+++ src/include/duckdb/function/create_sort_key.hpp
@@ -55,10 +55,6 @@ struct CreateSortKeyHelpers {
 	static void CreateSortKeyWithValidity(const Vector &input, Vector &result, const OrderModifiers &modifiers);
 	static void CreateSortKeyWithValidity(const Vector &input, Vector &result, const OrderModifiers &modifiers,
 	                                      idx_t count);
-	//! Encodes the *logical* value of a VARIANT into a byte-comparable sort key (used by the
-	//! variant_comparator collation). Unlike CreateSortKey this encoding is not reversible. NULLs are
-	//! propagated into the result validity.
-	static void CreateVariantSortKey(const Vector &input, idx_t count, OrderModifiers modifiers, Vector &result);
 };
 
 //! We don't add this function to the catalog, for internal use only
